fix: CommentService package restructure (change to edu.miu.cs544)
