+++ .idea/modules.xml
@@ -6,8 +6,9 @@
       <module fileurl="file://$PROJECT_DIR$/../Blog-it.iml" filepath="$PROJECT_DIR$/../Blog-it.iml" />
       <module fileurl="file://$PROJECT_DIR$/DiscoveryServer/DiscoveryServer.iml" filepath="$PROJECT_DIR$/DiscoveryServer/DiscoveryServer.iml" />
       <module fileurl="file://$PROJECT_DIR$/authentication-authorization-service/authentication-authorization-service.iml" filepath="$PROJECT_DIR$/authentication-authorization-service/authentication-authorization-service.iml" />
-      <module fileurl="file://$PROJECT_DIR$/blog/blog.iml" filepath="$PROJECT_DIR$/blog/blog.iml" />
+      <module fileurl="file://$PROJECT_DIR$/blog/blog-service.iml" filepath="$PROJECT_DIR$/blog/blog-service.iml" />
       <module fileurl="file://$PROJECT_DIR$/comments-service/comments-service.iml" filepath="$PROJECT_DIR$/comments-service/comments-service.iml" />
+      <module fileurl="file://$PROJECT_DIR$/user-service/user-service.iml" filepath="$PROJECT_DIR$/user-service/user-service.iml" />
     </modules>
   </component>
 </project>
