--- comments-service/src/main/java/com/example/blogpostservice/CommentServiceApplication.java
@@ -1,20 +1,0 @@
-package com.example.blogpostservice;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@SpringBootApplication
-@EnableJpaRepositories
-public class CommentServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CommentServiceApplication.class, args);
-    }
-
-    @Bean
-    public ModelMapper modelMapper(){return new ModelMapper();}
-
-}
