+++ comments-service/src/main/java/edu/miu/cs544/CommentsApp.java
@@ -1,11 +1,20 @@
 package edu.miu.cs544;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
+@EnableJpaRepositories
 public class CommentsApp {
+
     public static void main(String[] args) {
         SpringApplication.run(CommentsApp.class, args);
     }
+
+    @Bean
+    public ModelMapper modelMapper(){return new ModelMapper();}
+
 }
