+++ comments-service/src/main/java/edu/miu/cs544/dao/CommentDao.java
@@ -1,6 +1,6 @@
-package com.example.blogpostservice.dao;
+package edu.miu.cs544.dao;
 
-import com.example.blogpostservice.model.Comment;
+import edu.miu.cs544.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
