+++ comments-service/src/main/java/edu/miu/cs544/model/Auditable.java
@@ -1,0 +1,24 @@
+package edu.miu.cs544.model;
+
+import lombok.Data;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EntityListeners;
+import javax.persistence.MappedSuperclass;
+
+@Data
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public class Auditable {
+    @Column(name = "created_date", nullable = false, updatable = false)
+    @CreatedDate
+    private long createdDate;
+
+    @Column(name = "modified_date")
+    @LastModifiedDate
+    private long modifiedDate;
+}
