+++ comments-service/src/main/java/edu/miu/cs544/model/Comment.java
@@ -1,19 +1,19 @@
-package com.example.blogpostservice.model;
+package edu.miu.cs544.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 
 import javax.persistence.*;
-import java.io.Serializable;
 
 @Entity
 @Data
 @JsonIgnoreProperties(value = "hibernateLazyInitializer")
-public class Comment {
+public class Comment extends Auditable{
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
     @Lob
     private String content;
+
     private Long post;
     private Long user;
 }
