+++ comments-service/src/main/java/edu/miu/cs544/service/CommentService.java
@@ -1,7 +1,7 @@
-package com.example.blogpostservice.service;
+package edu.miu.cs544.service;
 
-import com.example.blogpostservice.dto.CommentRequest;
-import com.example.blogpostservice.model.Comment;
+import edu.miu.cs544.dto.CommentRequest;
+import edu.miu.cs544.model.Comment;
 
 import java.util.List;
