+++ comments-service/src/test/java/edu/miu/cs544/BlogCommentServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.blogpostservice;
+package edu.miu.cs544;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
